chore: Update appsettings.json for variable substitution
The code changes involve updating the appsettings.json file to use variable substitution for the "ServerURL" property. This allows for dynamic configuration of the server URL during deployment. The variable substitution is performed using the `microsoft/variable-substitution` GitHub Action. The value of the "ServerURL" property is now set to the value of the `PRODUCTION_SERVER_URL` environment variable.